+++ nilearn/decoding/_objective_functions.py
