+++ nilearn/decoding/proximal_operators.py
@@ -7,7 +7,7 @@ from math import sqrt
 
 import numpy as np
 
-from .objective_functions import _div_id, _gradient_id, tv_l1_from_gradient
+from ._objective_functions import _div_id, _gradient_id, tv_l1_from_gradient
 
 
 def _prox_l1(y, alpha, copy=True):
