+++ nilearn/decoding/tests/test_tv.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from nilearn.decoding.objective_functions import _gradient_id, _squared_loss
+from nilearn.decoding._objective_functions import _gradient_id, _squared_loss
 from nilearn.decoding.space_net_solvers import (
     _tvl1_objective,
     _tvl1_objective_from_gradient,
